Address chat context review feedback
