+++ apps/ui/src/ui-desks/chats/search.ts
@@ -1,4 +1,4 @@
-export interface DeskChatsSearch {
+export interface ChatsSearch {
 	chats?: boolean;
 	newChat?: number;
 }
@@ -12,7 +12,7 @@ function parseNewChatSearch( value: unknown ) {
 	return Number.isFinite( parsed ) && parsed > 0 ? parsed : undefined;
 }
 
-export function validateDeskChatsSearch( search: Record< string, unknown > ): DeskChatsSearch {
+export function validateChatsSearch( search: Record< string, unknown > ): ChatsSearch {
 	return {
 		chats: parseChatsSearch( search.chats ) || undefined,
 		newChat: parseNewChatSearch( search.newChat ),
