+++ apps/ui/src/ui-desks/chats/selection-chat-dialog/style.module.css
@@ -1,0 +1,172 @@
+.backdrop {
+	position: fixed;
+	inset: 0;
+	z-index: 700;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	background: rgba(15, 23, 42, 0.18);
+	-webkit-backdrop-filter: blur(8px);
+	backdrop-filter: blur(8px);
+	pointer-events: auto;
+	animation: fadeIn 200ms ease;
+}
+
+.dialog {
+	width: min(560px, calc(100vw - 48px));
+	display: flex;
+	flex-direction: column;
+	gap: 16px;
+	box-sizing: border-box;
+	padding: 20px;
+	border: 1px solid rgba(255, 255, 255, 0.7);
+	border-radius: var(--ui-desks-radius-panel, 22px);
+	corner-shape: var(--ui-desks-corner-shape, superellipse(1.42));
+	-webkit-corner-shape: var(--ui-desks-corner-shape, superellipse(1.42));
+	background: rgba(255, 255, 255, 0.62);
+	box-shadow:
+		0 1px 2px rgba(15, 23, 42, 0.04),
+		0 18px 44px rgba(15, 23, 42, 0.1);
+	color: var(--ui-desks-text, #14171a);
+	-webkit-backdrop-filter: saturate(180%) blur(28px);
+	backdrop-filter: saturate(180%) blur(28px);
+	animation: rise 280ms cubic-bezier(0.32, 0.72, 0.24, 1);
+}
+
+.promptRow {
+	display: flex;
+	align-items: flex-end;
+	gap: 8px;
+}
+
+.promptInput {
+	flex: 1;
+	min-width: 0;
+	max-height: 220px;
+	box-sizing: border-box;
+	margin: 0;
+	padding: 8px 4px;
+	border: 0;
+	background: transparent;
+	color: var(--ui-desks-text, #14171a);
+	font: inherit;
+	font-size: 15px;
+	line-height: 1.45;
+	letter-spacing: 0;
+	outline: none;
+	resize: none;
+	overflow-y: auto;
+}
+
+.promptInput::placeholder {
+	color: var(--ui-desks-muted, #6b7280);
+}
+
+.sendButton {
+	--control-button-background: #3858e9;
+	--control-button-foreground: #fff;
+
+	width: 40px;
+	min-width: 40px;
+	height: 40px;
+	flex: 0 0 auto;
+	border-radius: 12px;
+	box-shadow: 0 6px 18px rgba(56, 88, 233, 0.45);
+	transition:
+		opacity 200ms ease,
+		transform 280ms cubic-bezier(0.32, 0.72, 0.24, 1),
+		box-shadow 160ms ease;
+}
+
+.sendButton:hover:not(:disabled) {
+	--control-button-background: #3858e9;
+
+	box-shadow: 0 8px 22px rgba(56, 88, 233, 0.55);
+	transform: translateY(-1px);
+}
+
+.sendButton.sendButton:disabled {
+	opacity: 0;
+	pointer-events: none;
+	box-shadow: none;
+	transform: scale(0.6);
+}
+
+.sendButton svg {
+	fill: #fff;
+}
+
+.thumbnails {
+	display: flex;
+	flex-wrap: wrap;
+	align-items: flex-start;
+	gap: 8px;
+}
+
+.thumbnail {
+	position: relative;
+	flex: 0 0 auto;
+	overflow: visible;
+	pointer-events: none;
+}
+
+.thumbnailInner {
+	position: absolute;
+	top: 0;
+	left: 0;
+	transform-origin: top left;
+	pointer-events: none;
+}
+
+.moreThumbnail {
+	width: 72px;
+	height: 96px;
+	flex: 0 0 auto;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	box-sizing: border-box;
+	padding: 8px;
+	border: 1px solid rgba(255, 255, 255, 0.72);
+	border-radius: 10px;
+	corner-shape: var(--ui-desks-corner-shape, superellipse(1.42));
+	-webkit-corner-shape: var(--ui-desks-corner-shape, superellipse(1.42));
+	background: rgba(255, 255, 255, 0.58);
+	box-shadow:
+		0 1px 2px rgba(15, 23, 42, 0.04),
+		0 14px 30px rgba(15, 23, 42, 0.08);
+	color: var(--ui-desks-text, #14171a);
+	font-size: 12px;
+	line-height: 1.2;
+	font-weight: 600;
+	letter-spacing: 0;
+	text-align: center;
+}
+
+.error {
+	color: #b32d2e;
+	font-size: 12px;
+	line-height: 16px;
+}
+
+@keyframes fadeIn {
+	from {
+		opacity: 0;
+	}
+
+	to {
+		opacity: 1;
+	}
+}
+
+@keyframes rise {
+	from {
+		opacity: 0;
+		transform: translateY(8px) scale(0.98);
+	}
+
+	to {
+		opacity: 1;
+		transform: translateY(0) scale(1);
+	}
+}
