+++ apps/ui/src/ui-desks/router/root.tsx
@@ -1,7 +1,7 @@
 import { createRootRoute, Outlet } from '@tanstack/react-router';
-import { validateDeskChatsSearch } from '../chats/search';
+import { validateChatsSearch } from '../chats/search';
 
 export const desksRootRoute = createRootRoute< unknown >( {
-	validateSearch: validateDeskChatsSearch,
+	validateSearch: validateChatsSearch,
 	component: () => <Outlet />,
 } );
