+++ apps/ui/src/ui-desks/widgets/toolbar.module.css
@@ -29,199 +29,3 @@
 	line-height: 36px;
 	white-space: nowrap;
 }
-
-.chatToolbarButton {
-	--control-button-background: #3858e9;
-	--control-button-foreground: #fff;
-
-	box-shadow: 0 6px 18px rgba(56, 88, 233, 0.45);
-	transition:
-		background 140ms ease,
-		box-shadow 160ms ease,
-		transform 160ms ease;
-}
-
-.chatToolbarButton:hover:not(:disabled) {
-	--control-button-background: #3858e9;
-
-	box-shadow: 0 8px 22px rgba(56, 88, 233, 0.55);
-	transform: translateY(-1px);
-}
-
-.chatToolbarButton svg {
-	fill: #fff;
-}
-
-.chatDialogBackdrop {
-	position: fixed;
-	inset: 0;
-	z-index: 40;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding: 36px 38px;
-	background: rgba(15, 23, 42, 0.08);
-	-webkit-backdrop-filter: blur(12px) saturate(140%);
-	backdrop-filter: blur(12px) saturate(140%);
-	pointer-events: auto;
-}
-
-.chatDialog {
-	position: relative;
-	width: min(680px, calc(100vw - 76px));
-	min-height: min(228px, calc(100vh - 72px));
-	display: flex;
-	flex-direction: column;
-	box-sizing: border-box;
-	padding: 34px 88px 28px 34px;
-	border: 1px solid rgba(255, 255, 255, 0.78);
-	border-radius: 30px;
-	corner-shape: var(--ui-desks-corner-shape, superellipse(1.42));
-	-webkit-corner-shape: var(--ui-desks-corner-shape, superellipse(1.42));
-	background: rgba(236, 239, 240, 0.72);
-	box-shadow:
-		0 1px 0 rgba(255, 255, 255, 0.82) inset,
-		0 24px 72px rgba(15, 23, 42, 0.16);
-	color: var(--ui-desks-text, #14171a);
-	-webkit-backdrop-filter: saturate(150%) blur(34px);
-	backdrop-filter: saturate(150%) blur(34px);
-}
-
-.chatPromptInput {
-	width: min(520px, 100%);
-	min-height: 30px;
-	max-height: 90px;
-	box-sizing: border-box;
-	margin: 0;
-	padding: 0;
-	border: 0;
-	background: transparent;
-	color: var(--ui-desks-text, #14171a);
-	font: inherit;
-	font-size: 18px;
-	line-height: 1.25;
-	letter-spacing: 0;
-	outline: none;
-	resize: none;
-}
-
-.chatPromptInput::placeholder {
-	color: rgba(20, 23, 26, 0.38);
-}
-
-.chatSendButton {
-	--control-button-background: #3858e9;
-	--control-button-foreground: #fff;
-
-	position: absolute;
-	top: 26px;
-	right: 26px;
-	width: 52px;
-	min-width: 52px;
-	height: 52px;
-	border-radius: 16px;
-	box-shadow:
-		0 20px 36px rgba(56, 88, 233, 0.28),
-		0 10px 18px rgba(15, 23, 42, 0.12);
-}
-
-.chatSendButton:hover:not(:disabled) {
-	--control-button-background: #3151e7;
-}
-
-.chatSendButton:active:not(:disabled) {
-	transform: translateY(1px);
-}
-
-.chatVignettes {
-	display: flex;
-	flex-wrap: wrap;
-	align-items: flex-start;
-	gap: 14px;
-	margin-top: 24px;
-	max-height: calc(100vh - 184px);
-	overflow: auto;
-}
-
-.chatVignette {
-	position: relative;
-	flex: 0 0 auto;
-	overflow: visible;
-	pointer-events: none;
-}
-
-.chatVignetteInner {
-	position: absolute;
-	top: 0;
-	left: 0;
-	transform-origin: top left;
-	pointer-events: none;
-}
-
-.chatMoreVignette {
-	width: 88px;
-	height: 88px;
-	flex: 0 0 auto;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	box-sizing: border-box;
-	border: 1px solid rgba(255, 255, 255, 0.72);
-	border-radius: 16px;
-	background: rgba(255, 255, 255, 0.58);
-	box-shadow:
-		0 1px 2px rgba(15, 23, 42, 0.04),
-		0 14px 30px rgba(15, 23, 42, 0.08);
-	color: var(--ui-desks-text, #14171a);
-	font-size: 14px;
-	line-height: 1.2;
-	font-weight: 600;
-	letter-spacing: 0;
-}
-
-.chatDialogError {
-	margin-top: 18px;
-	color: #b32d2e;
-	font-size: 12px;
-	line-height: 16px;
-}
-
-@media (max-width: 700px) {
-	.chatDialogBackdrop {
-		padding: 16px;
-	}
-
-	.chatDialog {
-		width: calc(100vw - 32px);
-		min-height: min(220px, calc(100vh - 32px));
-		padding: 28px 22px 24px;
-		border-radius: 28px;
-	}
-
-	.chatPromptInput {
-		width: calc(100% - 72px);
-		font-size: 18px;
-		min-height: 30px;
-	}
-
-	.chatSendButton {
-		top: 22px;
-		right: 22px;
-		width: 48px;
-		min-width: 48px;
-		height: 48px;
-		border-radius: 16px;
-	}
-
-	.chatVignettes {
-		margin-top: 22px;
-		gap: 12px;
-	}
-
-	.chatMoreVignette {
-		width: 80px;
-		height: 80px;
-		border-radius: 14px;
-		font-size: 13px;
-	}
-}
